Organize devtools
