+++ devtools/browserslist-config/browserslist-config.cjs
@@ -1,0 +1,3 @@
+module.exports = [
+  'node 22',
+];
