+++ devtools/rollup/src/main.ts
@@ -1,0 +1,1 @@
+export { createRollupConfig } from './rollup.js';
