+++ packages/jsx-scanner/babel.config.json
@@ -1,3 +1,3 @@
 {
-  "extends": "@coscan/devtools/babel.config.json"
+  "extends": "@coscan/babel-config"
 }
