+++ packages/jsx-scanner/rollup.config.js
@@ -1,4 +1,4 @@
-import { createRollupConfig } from '@coscan/devtools';
+import { createRollupConfig } from '@coscan/rollup';
 import pkg from './package.json' with { type: 'json' };
 
 export default createRollupConfig(pkg);
