+++ packages/jsx-scanner/tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "@coscan/devtools/tsconfig.json"
+  "extends": "@coscan/tsconfig"
 }
