--- utilities/devtools/src/main.ts
@@ -1,1 +1,0 @@
-export { createRollupConfig } from './rollup.ts';
